test(l1): move eth_simulateV1 integration tests to test/tests/rpc/
Move 13 integration tests from the simulate module to
test/tests/rpc/simulate_tests.rs, keeping 14 unit tests (parsing,
serialization, header construction) in simulate.rs.

Made eth::simulate module and eth module pub for external test access.
