+++ crates/networking/rpc/eth/mod.rs
@@ -9,4 +9,4 @@ pub(crate) mod transaction;
 pub(crate) mod gas_price;
 pub(crate) mod gas_tip_estimator;
 pub(crate) mod max_priority_fee;
-pub(crate) mod simulate;
+pub mod simulate;
