+++ test/tests/rpc/mod.rs
@@ -1,1 +1,2 @@
 mod client_version_tests;
+mod simulate_tests;
